Replaced BigDecimal with Integer in JournalEntryService.
Also in PotManagerService.

+++ impwrme2/src/main/java/com/impwrme2/service/engine/IResourceEngine.java
@@ -1,6 +1,5 @@
 package com.impwrme2.service.engine;
 
-import java.math.BigDecimal;
 import java.time.YearMonth;
 
 import com.impwrme2.model.resource.Resource;
@@ -8,9 +7,9 @@ import com.impwrme2.model.resource.Resource;
 public interface IResourceEngine {
 
 	public Resource getResource();
-	public BigDecimal getBalanceLiquidLegalMax(final YearMonth yearMonth);
-	public BigDecimal getBalanceLiquidLegalMin(final YearMonth yearMonth);
-	public BigDecimal getBalanceLiquidPreferredMax(final YearMonth yearMonth);
-	public BigDecimal getBalanceLiquidPreferredMin(final YearMonth yearMonth);
+	public Integer getBalanceLiquidLegalMax(final YearMonth yearMonth);
+	public Integer getBalanceLiquidLegalMin(final YearMonth yearMonth);
+	public Integer getBalanceLiquidPreferredMax(final YearMonth yearMonth);
+	public Integer getBalanceLiquidPreferredMin(final YearMonth yearMonth);
 //	public List<JournalEntry> generateJournalEntriesForExpensesIncomeAndWithdrawals(final PotManagerService potManager);
 }
